refactor: LottoNumbers 관련 로직을 Lotto 로 이동
